lib: monkey: sync changes (details below)
 - 041a3570 core: build: fix memrchr() validation
 - f7037716 core: rconf: change buffer from static 1kb to dynamic 4kb
 - ac6c8514 api: test: fix test case formatter type
 - 347a9a5a server: thread: comment out unused code
 - f0fdabb7 core: declare built-in memrchr
 - d5c36947 server: cast type to format pid_t
 - f145e816 core: utils: cast type to format pid_t
 - 5bd56d70 server: build: on Solaris link to socket and nsl
 - a913d53c core: validate if memrchr() exists

+++ lib/monkey/mk_server/monkey.c
@@ -31,7 +31,7 @@ void mk_server_info(struct mk_server *server)
     struct mk_plugin *p;
     struct mk_config_listener *l;
 
-    printf(MK_BANNER_ENTRY "Process ID is %i\n", getpid());
+    printf(MK_BANNER_ENTRY "Process ID is %ld\n", (long) getpid());
     mk_list_foreach(head, &server->listeners) {
         l = mk_list_entry(head, struct mk_config_listener, _head);
         printf(MK_BANNER_ENTRY "Server listening on %s:%s\n",
